Load all resolutions upon model creation.
If we don't do so then it becomes impossible to load all of them since m_aspectRatio
will be empty and the only way to load all them is to set the aspectRatio property
to an empty string.
The new property value and the stored one will always be equal in such case.