Infrastructure for filters.
This consists of:
A GStreamer element to duplicate buffers
A wrapper that creates a bin utilizing the above mentioned element.

On some devices like N9, memory is frame buffer memory and access to such
memory is costly since it's not cached by the CPU. What we do is we copy it
to "normal memory" and pass that to any filters that want to inspect the buffer.