Reworked DoneHandler to avoid race conditions
If we start capturing the 2nd image before the first gets saved then we
accidentally try to rename the 1st image to the 2nd image.

We now only try to rename if we have a temporarily file name and a mode file name.
Since video mode only sets tempFileName then we should be fine as we don't really
allow recording a new video before the old one gets saved.

For images, we simply use the filename from the structure for the saved() signal
and fall back to the mode fileName if it's empty without overwriting the members
of QtCamModePrivate