Race condition when QtCamVideo::stop() gets called multiple times.
camerabin2 sends the video-done message before setting idle to true.

If we get a chance to call QtCamVideoMode::stop() for a second time before
camerabin2 gets a chance to reset its idle property to true then we deadlock.

Introducing a boolean value to indicate that we have received the video-done message
fixes the issue. said boolean is always accessed while mutex is locked to avoid
any race conditions.