Don't disable GPS when we are obscured by location-ui or conndlgs
Achieved by adding a PlatformQuirks QML component which takes care of
forcing camera to be on.

On Harmattan it forces on when we are obscured by location-ui (for positioning terms)
or conndlgs (for connection establishment).

As a side effect, we now prevent display dimming while playing a video in post capture