Properly restart mtheme daemon in postinst script
It's a hack but here is what we do:
- Have a script installed in /usr/bin
- Request setuid user bit for that script via aegis manifest
- Execute the script in our postinst script

Since the script is setuid user, it will be executed as the "user" user.
The script will send a sig HUP to mtheme daemon which will cause it
to reload the theme.

It's ugly but it works.
Most likely we don't need the setgid root bit but I am too lazy to try again.