don't handle zoom via camerabin
the camera source can update its max-zoom property and zoom GParamSpec depending
on the preview size. camerabin will update its max-zoom property but will not
update the zoom GParamSpec which can lead to failure to set the zoom property if
the value we are setting is more than the maximum value defined in the GParamSpec
The value we are setting is valid but camerabin is using an old spec definition and
thus failing us.