Added a start() virtual method to QtCamViewfinder and to the renderers
SailfishOS (an possibly others) send a paint event after we reset the renderer
which causes the renderer to try to set the egl display on the sink. Since this
happens after we get minimized, the viewfinder fails to render when it gets
started again. A workaround is to return from ::paint() as long as the viewfinder
renderer has not been started yet.