Added ImageSettings::bestResolution()
This method will check that a given aspect ratio has a given resolution.
If not then it will pick the first resolution and return it.

The aim is to have a valid resolution when we switch the aspect ratio

Some aspect ratioss can have different resolution names (like N950 4:3
which has a huge resolution corresponding to its 12 MP sensor)