declarativeqtcamera: Account for device deletion while capabilities are still using it
We will delete the wrapped capability when the device gets deleted otherwise we will be
accessing garbage memory.

It should not be a big deal because we will recreate our capabilities very soon.

This fixes a rare crash that was happening when switching between front and back cameras
because we were manipulating Roi during such transition

Also created a minimal declarative camera test