Add hidden resolutions.
These are resolutions that the camera module supports but are not visible to the user.
They can be used for doing image analysis (Panorama for example). and can be used like
other resolutions.

They are not supported because they might be too low for capturing for example or resulting
images are not so good but we only care about the viewfinder resolution since we will
use it to do some analysis.